refactor!: rename purify module to functional
BREAKING CHANGE: Adjective usage is more proper than noun usage. Also,
this module will eventually introduce more definitions which are not
coming from purify.ts.
