+++ src/-error.ts
@@ -5,13 +5,13 @@
  * @module
  */
 
-import { Maybe } from './-purify';
+import { Maybe } from './-functional';
 
 /**
  * Type definition for custom, non-throwable errors.
  *
- * This is a useful type for the {@link purify!Left} of an {@link purify!Either}
- * value.
+ * This is a useful type for the {@link functional!Left} of an
+ * {@link functional!Either} value.
  */
 export interface CustomError {
   msg: string;
@@ -30,8 +30,8 @@ export function customError(msg: string, err?: Error): CustomError {
 }
 
 /**
- * Convenience function for throwing errors indicating that a definition is
- * not implemented yet.
+ * Convenience function for throwing errors indicating that a definition is not
+ * implemented yet.
  *
  * This is useful during development. Production code should not contain calls
  * to this function.
