+++ src/-functional.ts
@@ -1,0 +1,8 @@
+/**
+ * This module re-exports purify.ts definitions and provides further functional
+ * definitions.
+ *
+ * @module
+ */
+
+export * from 'purify-ts';
