--- src/-purify.ts
@@ -1,7 +1,0 @@
-/**
- * This module re-exports purify.ts definitions.
- *
- * @module
- */
-
-export * from 'purify-ts';
